+++ src/routers/oauth2.ts
@@ -3,6 +3,7 @@ import validationErrorHandler from '../middleware/validationErrorHandler'
 import * as validators from '../middleware/validators'
 import setHeader from '../middleware/setHeader'
 import { tokenRoute } from '../routes/oauth2/token'
+import { authorizationCode } from '../config/gamma'
 
 function createOAuth2Router(): Router {
     const router = Router()
@@ -14,6 +15,10 @@ function createOAuth2Router(): Router {
         validationErrorHandler,
         tokenRoute()
     )
+
+    router.get('/authorize', (_req, res) => {
+        res.redirect(authorizationCode.authorizeUrl())
+    })
 
     return router
 }